+++ docs/HISTORY.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+'0.1'
+-----
+
+ - Initial content type and template
+ - Full text index all fields
+
 '0.1' - Unreleased
 ---------------------